+++ activestorage/lib/active_storage/attached/changes/detach_many.rb
@@ -10,7 +10,9 @@ module ActiveStorage
 
     def detach
       if attachments.any?
+        blob_ids = attachments.filter_map { |a| a.blob_id if a.persisted? }
         attachments.delete_all if attachments.respond_to?(:delete_all)
+        ActiveStorage::Blob.where(id: blob_ids).touch_all if blob_ids.any?
         record.attachment_changes.delete(name)
       end
     end