+++ activestorage/test/test_helper.rb
@@ -158,9 +158,11 @@ class User < ActiveRecord::Base
   has_one_attached :avatar_with_immediate_analysis, analyze: :immediately
   has_one_attached :avatar_with_later_analysis, analyze: :later
   has_one_attached :avatar_with_lazy_analysis, analyze: :lazily
+  has_one_attached :profile_photo, dependent: :detach
 
   has_many_attached :highlights
   has_many_attached :vlogs, dependent: false, service: :local
+  has_many_attached :documents, dependent: :detach
   has_many_attached :highlights_with_variants do |attachable|
     attachable.variant :thumb, resize_to_limit: [100, 100]
   end